Fix compiler warning.

+++ trunk/src/realjsformatter.cpp
@@ -127,9 +127,9 @@ void RealJSFormatter::PrintAdditionalDebug(string& strDebugOutput)
 	strDebugOutput.append(buf);
 }
 
-int RealJSFormatter::GetFormattedLine(unsigned int originalLine)
+int RealJSFormatter::GetFormattedLine(int originalLine)
 {
-	if(originalLine <= 0 || m_lineFormattedVec.size() <= originalLine)
+	if(originalLine <= 0 || m_lineFormattedVec.size() <= (size_t)originalLine)
 		return -1;
 
 	for(int l = originalLine; l > 0; --l)
@@ -223,9 +223,9 @@ void RealJSFormatter::PutString(const Token& token)
 		else
 		{
 			m_lineBuffer += token.code[i];
-			int tokenLine = token.line;
+			int tokenLine = (int)token.line;
 			if(tokenLine != -1)
-				m_lineWaitVec.push_back(token.line);
+				m_lineWaitVec.push_back((int)token.line);
 		}
 	}
 }
@@ -253,8 +253,8 @@ void RealJSFormatter::PutLineBuffer()
 			break;
 		}
 
-		unsigned int oldLine = m_lineWaitVec[i];
-		if(oldLine >= m_lineFormattedVec.size())
+		int oldLine = m_lineWaitVec[i];
+		if((size_t)oldLine >= m_lineFormattedVec.size())
 		{
 			m_lineFormattedVec.resize(m_lineFormattedVec.size()*2, -1);
 			continue;
